requirements.txt
pytest
requests
python-dotenv
allure-pytest
pytest-order